% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compdag-package.R
\docType{package}
\name{compdag-package}
\alias{compdag-package}
\alias{compdag}
\title{The 'compdag' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.23. https://mc-stan.org
}
