% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compdag.R
\name{compdag}
\alias{compdag}
\title{compdag}
\usage{
compdag(x, y, paramx, paramy)
}
\arguments{
\item{x}{data of the first community}

\item{y}{data of the second scommunity}

\item{paramx}{the fixed mu for the first community}

\item{paramy}{the fixed mu for the second community}
}
\value{
cd=1 indicates that the community-level causal direction is x to y;
cd=-1 indicates that the community-level causal direction is y to x;
cd = 0 indicate that there is no causal relationship between x and y.
mEst is the estimated microbe-level causal effects.
signalEst is the estimated community-level causal effect.
x2y is the looic of the model x to y;
y2x is the looic of the model y to x;
x0y is the looic of the model that there is no casual relationship.
}
\description{
compdag
}
